chore: tidy up
move request and events to utils, as shared structures
reuse the InputRequest in events
